+++ CHANGES.rst
@@ -15,6 +15,7 @@ Raster Caster changelog
 - Removed the debug code that wrote the triangulation to a GeoPackage.
 - Added validation on input layers.
 - Elevation point layer is now a Point layer, not PointZ.
+- input files are now properly unlocked in case of validation errors.
 
 
 0.2 (2026-09-21)