+++ CHANGES.rst
@@ -20,6 +20,7 @@ Raster Caster changelog
 - Inputs are now layers instead of geopackage.
 - Generate geopackage: no longer able to save to temporary file.
 - Set styling to output raster.
+- Added in_polygon_only support
 
 0.2 (2026-09-21)
 ----------------